Run PushTx and PandoTx into the same process
