+++ pm2.config.cjs.example
@@ -45,19 +45,6 @@ module.exports = {
             kill_timeout: 3000,
         },
         {
-            name: `Samourai Dojo - PandoTX (${NAMESPACE})`,
-            namespace: NAMESPACE,
-            script: './index-pandotx.js',
-            cwd: path.join(__dirname, 'pushtx'),
-            interpreter: INTERPRETER,
-            out_file: './output-pandotx-2022-01-01.log',
-            error_file: './error-pandotx-2022-01-01.log',
-            wait_ready: true,
-            stop_exit_codes: 0,
-            listen_timeout: 5000,
-            kill_timeout: 3000,
-        },
-        {
             name: `Samourai Dojo - Tracker (${NAMESPACE})`,
             namespace: NAMESPACE,
             script: './index.js',
