+++ pushtx/index.js
@@ -3,16 +3,43 @@
  * Copyright © 2019 – Katana Cryptographic Ltd. All Rights Reserved.
  */
 
+import { RPC } from 'soroban-client-nodejs'
 import Logger from '../lib/logger.js'
 import db from '../lib/db/mysql-db-wrapper.js'
 import { waitForBitcoindRpcApi } from '../lib/bitcoind-rpc/rpc-client.js'
 import network from '../lib/bitcoin/network.js'
 import keysFile from '../keys/index.js'
 import HttpServer from '../lib/http-server/http-server.js'
+import util from '../lib/util.js'
 import PushTxRestApi from './pushtx-rest-api.js'
 import pushTxProcessor from './pushtx-processor.js'
+import pandoTx from './pandotx-processor.js'
+
 
 const keys = keysFile[network.key]
+
+
+/**
+ * Check if the Soroban rpc api is ready to process requests
+ * @returns {Promise<void>}
+ */
+async function waitForSorobanRpcApi() {
+    let sorobanRpc = null
+    try {
+        const sorobanUrl = keys['pandoTx']['sorobanUrl']
+        const socks5ProxyUrl = (sorobanUrl.includes('.onion')) ?
+            keys['pandoTx']['socks5Proxy'] :
+            null
+        sorobanRpc = new RPC(sorobanUrl, socks5ProxyUrl)
+        const entries = await sorobanRpc.directoryList('test.start')
+    } catch (e) {
+        sorobanRpc = null
+        Logger.info('PandoTx : Soroban RPC API is still unreachable. New attempt in 20s.')
+        await util.delay(20000)
+        return await waitForSorobanRpcApi()
+    }
+}
+
 
 try {
     /**
@@ -24,6 +51,10 @@ try {
     // Wait for Bitcoind RPC API
     // being ready to process requests
     await waitForBitcoindRpcApi()
+
+    // Wait for Soroban RPC API
+    // being ready to process requests
+    await waitForSorobanRpcApi()
 
     // Initialize the db wrapper
     const dbConfig = {
@@ -53,11 +84,16 @@ try {
     // Start the http server
     httpServer.start()
 
+    // Start the processor
+    pandoTx.start()
+    Logger.info('PushTx : PandoTx processor started')
+
     // Signal that the process is ready
     process.send('ready')
 
     const exit = async () => {
         httpServer.stop()
+        pandoTx.stop()
         await db.disconnect()
         process.exit(0)
     }
